requirements.txt
Django==1.8
Pillow==2.8.1
baidupcsapi==0.3.4
image==1.3.4
pyasn1==0.1.7
requests==2.6.0
requests-toolbelt==0.4.0
rsa==3.1.4
wget==2.2
wsgiref==0.1.2
